README updates, moved objects
The SD card, GPS, and UART objects are moved to files for the tasks which use them.
GPS pins are changed (seemed to be wrong ones?) and a bug fixed where the GPS TXD (ESP32 RXD) pin was configured by MicroPython in a way that prevented reliable GPS data transmission.
MQTT topic name is configured using the configuration file site name.